Refactorizar la lógica de precios de productos para utilizar `BigDecimal` y garantizar la precisión, hacer opcional la descripción de la categoría y mejorar la configuración de CORS para permitir el método PATCH.

+++ foodstore-backend/src/main/java/com/tuuniversidad/foodstore/dto/producto/ProductoEdit.java
@@ -8,6 +8,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -18,7 +20,7 @@ public class ProductoEdit {
     private String nombre;
 
     @DecimalMin(value = "0.01", message = "El precio debe ser mayor a 0")
-    private Double precio;
+    private BigDecimal precio;
 
     private String descripcion;
